Extract TreeSearch from TreeViewHeader.
