+++ src/components/TreeView/TreeViewHeader.tsx
@@ -1,12 +1,10 @@
 import {
   Button,
-  Card,
   ImageList,
   ImageListItem,
-  TextField,
   Typography,
 } from "@material-ui/core";
-import React, { useCallback, useEffect, useState } from "react";
+import { useCallback, useEffect } from "react";
 import Bounce from "react-reveal/Bounce";
 import { Key } from "ts-key-enum";
 
@@ -21,12 +19,7 @@ export interface TreeHeaderProps {
 }
 
 export function TreeViewHeader(props: TreeHeaderProps) {
-  const {
-    getLeftBrother,
-    getRightBrother,
-    searchAndSelectDomain,
-    handleChange,
-  } = useTreeViewNavigation(props);
+  const { getLeftBrother, getRightBrother } = useTreeNavigation(props);
 
   return (
     <ImageList cols={9} gap={20} rowHeight={"auto"}>
@@ -43,38 +36,24 @@ export function TreeViewHeader(props: TreeHeaderProps) {
         ) : null}
       </ImageListItem>
       <ImageListItem cols={5}>
-        <Card>
-          <Bounce spy={props.bounceState} duration={2000}>
-            <Button
-              fullWidth
-              size="large"
-              color="primary"
-              variant="contained"
-              disabled={!props.currentDomain.parentDomain}
-              onClick={() => props.animate(props.currentDomain)}
-              id="current-domain"
-            >
-              <div style={{ textTransform: "capitalize" }}>
-                <Typography variant="overline">
-                  {props.currentDomain.id}
-                </Typography>
-                <Typography variant="h6">{props.currentDomain.name}</Typography>
-              </div>
-            </Button>
-          </Bounce>
-          <TextField
+        <Bounce spy={props.bounceState} duration={2000}>
+          <Button
             fullWidth
-            id="name"
-            label="Find a domain"
-            onKeyDown={searchAndSelectDomain}
-            onChange={handleChange}
-            margin="normal"
-            autoComplete="off"
-            inputProps={{
-              "data-testid": "testSearch",
-            }}
-          />
-        </Card>
+            size="large"
+            color="primary"
+            variant="contained"
+            disabled={!props.currentDomain.parentDomain}
+            onClick={() => props.animate(props.currentDomain)}
+            id="current-domain"
+          >
+            <div style={{ textTransform: "capitalize" }}>
+              <Typography variant="overline">
+                {props.currentDomain.id}
+              </Typography>
+              <Typography variant="h6">{props.currentDomain.name}</Typography>
+            </div>
+          </Button>
+        </Bounce>
       </ImageListItem>
       <ImageListItem cols={2}>
         {getRightBrother(props) ? (
@@ -93,8 +72,7 @@ export function TreeViewHeader(props: TreeHeaderProps) {
 }
 
 // exported for unit testing only
-export function useTreeViewNavigation(props: TreeHeaderProps) {
-  const [input, setInput] = useState(props.currentDomain.id);
+export function useTreeNavigation(props: TreeHeaderProps) {
   // Gets the domain 'navigationAmount' away from the currentDomain (negative to the left, positive to the right)
   function getBrotherDomain(
     navigationAmount: number,
@@ -145,93 +123,6 @@ export function useTreeViewNavigation(props: TreeHeaderProps) {
     [props]
   );
 
-  // Search for a semantic domain by number
-  function searchDomainByNumber(
-    parent: TreeSemanticDomain,
-    number: string
-  ): TreeSemanticDomain | undefined {
-    for (const domain of parent.subdomains)
-      if (domain.id === number) {
-        return domain;
-      }
-    if (parent.id === number) {
-      return parent;
-    }
-    return undefined;
-  }
-
-  // Searches for a semantic domain by name
-  function searchDomainByName(
-    domain: TreeSemanticDomain,
-    target: string
-  ): TreeSemanticDomain | undefined {
-    let check = (checkAgainst: TreeSemanticDomain | undefined) =>
-      checkAgainst && target.toLowerCase() === checkAgainst.name.toLowerCase();
-    if (check(domain)) {
-      return domain;
-    }
-    // If there are subdomains
-    if (domain.subdomains.length > 0) {
-      let tempDomain: TreeSemanticDomain | undefined;
-      for (const sub of domain.subdomains) {
-        tempDomain = searchDomainByName(sub, target);
-        if (check(tempDomain)) {
-          return tempDomain;
-        }
-      }
-    }
-    return undefined;
-  }
-
-  // Dispatch the search for a specified domain, and switches to it if it exists
-  function searchAndSelectDomain(event: React.KeyboardEvent) {
-    // stopPropagation() prevents keystrokes from reaching ReviewEntries,
-    // but requires the search function be called onKeyDown
-    if (event.stopPropagation) {
-      event.stopPropagation();
-    }
-    event.bubbles = false;
-
-    if (event.key === Key.Enter) {
-      event.preventDefault();
-      // Find parent domain
-      let parent: TreeSemanticDomain | undefined = props.currentDomain;
-      while (parent.parentDomain !== undefined) {
-        parent = parent.parentDomain;
-      }
-
-      // Search for domain
-      if (!isNaN(parseInt(input))) {
-        let i: number = 0;
-        while (parent) {
-          parent = searchDomainByNumber(parent, input.slice(0, i * 2 + 1));
-          if (parent && parent.id === input) {
-            props.animate(parent);
-            props.bounce();
-            setInput("");
-            (event.target as any).value = "";
-            break;
-          } else if (parent && parent.subdomains.length === 0) {
-            break;
-          }
-          i++;
-        }
-      } else {
-        parent = searchDomainByName(parent, input);
-        if (parent) {
-          props.animate(parent);
-          props.bounce();
-          setInput("");
-          (event.target as any).value = "";
-        }
-      }
-    }
-  }
-
-  // Change the input on typing
-  function handleChange(event: React.ChangeEvent<HTMLTextAreaElement>) {
-    setInput(event.target.value);
-  }
   // Add event listeners
   useEffect(() => {
     window.addEventListener("keydown", navigateDomainArrowKeys);
@@ -244,7 +135,5 @@ export function useTreeViewNavigation(props: TreeHeaderProps) {
   return {
     getRightBrother,
     getLeftBrother,
-    searchAndSelectDomain,
-    handleChange,
   };
 }
