+++ src/components/TreeView/tests/TreeViewHeader.test.tsx
@@ -1,20 +1,17 @@
-import { render, screen } from "@testing-library/react";
-import { renderHook, act } from "@testing-library/react-hooks";
-import userEvent from "@testing-library/user-event";
-import React from "react";
+import { render } from "@testing-library/react";
+import { renderHook } from "@testing-library/react-hooks";
 import { Key } from "ts-key-enum";
 
 import MockDomain from "components/TreeView/tests/MockSemanticDomain";
 import {
   TreeViewHeader,
   TreeHeaderProps,
-  useTreeViewNavigation,
+  useTreeNavigation,
 } from "components/TreeView/TreeViewHeader";
 
 // Handles
 const MOCK_ANIMATE = jest.fn();
 const MOCK_BOUNCE = jest.fn();
-const MOCK_STOP_PROP = jest.fn();
 const testProps: TreeHeaderProps = {
   animate: MOCK_ANIMATE,
   currentDomain: MockDomain,
@@ -41,160 +38,9 @@ beforeEach(() => {
 });
 
 describe("TreeViewHeader", () => {
-  describe("searchAndSelectDomain", () => {
-    function setupSimulatedInputTest(input: string) {
-      // Simulate the user typing a string
-      const simulatedInput = {
-        target: { value: input },
-      } as React.ChangeEvent<HTMLTextAreaElement>;
-
-      const keyboardTarget = new EventTarget();
-      // Simulate the user typing the enter key
-      const simulatedEnterKey: Partial<React.KeyboardEvent> = {
-        bubbles: true,
-        key: Key.Enter,
-        preventDefault: jest.fn(),
-        target: keyboardTarget,
-        stopPropagation: MOCK_STOP_PROP,
-      };
-
-      return { simulatedInput, simulatedEnterKey };
-    }
-
-    it("switches semantic domain if given number found", () => {
-      const TEST: string = "1.0";
-
-      const { result } = renderHook(() => useTreeViewNavigation(testProps));
-
-      // Simulate the user typing
-      const { simulatedInput, simulatedEnterKey } =
-        setupSimulatedInputTest(TEST);
-
-      // When testing hooks any call that results in a state change needs to be wrapped in
-      // an act call to avoid warnings and make sure the state change is complete before we test
-      // for the results
-      act(() => result.current.handleChange(simulatedInput));
-      act(() =>
-        result.current.searchAndSelectDomain(
-          simulatedEnterKey as React.KeyboardEvent
-        )
-      );
-
-      expect(MOCK_STOP_PROP).toHaveBeenCalled();
-      expect(MOCK_BOUNCE).toHaveBeenCalled();
-      expect(MOCK_ANIMATE).toHaveBeenCalledWith(MockDomain.subdomains[0]);
-    });
-
-    it("does not switch semantic domain if given number not found", () => {
-      const TEST: string = "10";
-
-      const { result } = renderHook(() => useTreeViewNavigation(testProps));
-
-      // Simulate the user typing
-      const { simulatedInput, simulatedEnterKey } =
-        setupSimulatedInputTest(TEST);
-
-      // When testing hooks any call that results in a state change needs to be wrapped in
-      // an act call to avoid warnings and make sure the state change is complete before we test
-      // for the results
-      act(() => result.current.handleChange(simulatedInput));
-      act(() =>
-        result.current.searchAndSelectDomain(
-          simulatedEnterKey as React.KeyboardEvent
-        )
-      );
-
-      expect(MOCK_ANIMATE).toHaveBeenCalledTimes(0);
-    });
-
-    it("does not switch semantic domain on realistic but non-existent subdomain", () => {
-      const TEST: string = "1.2.1.1.1.1";
-
-      const { result } = renderHook(() => useTreeViewNavigation(testProps));
-
-      // Simulate the user typing
-      const { simulatedInput, simulatedEnterKey } =
-        setupSimulatedInputTest(TEST);
-
-      // When testing hooks any call that results in a state change needs to be wrapped in
-      // an act call to avoid warnings and make sure the state change is complete before we test
-      // for the results
-      act(() => result.current.handleChange(simulatedInput));
-      act(() =>
-        result.current.searchAndSelectDomain(
-          simulatedEnterKey as React.KeyboardEvent
-        )
-      );
-
-      expect(MOCK_ANIMATE).toHaveBeenCalledTimes(0);
-    });
-
-    it("switches on a length 5 number", () => {
-      const leafNode =
-        MockDomain.subdomains[2].subdomains[0].subdomains[0].subdomains[0];
-
-      const { result } = renderHook(() => useTreeViewNavigation(testProps));
-
-      // Simulate the user typing the leafNode.id
-      const { simulatedInput, simulatedEnterKey } = setupSimulatedInputTest(
-        leafNode.id
-      );
-
-      // When testing hooks any call that results in a state change needs to be wrapped in
-      // an act call to avoid warnings and make sure the state change is complete before we test
-      // for the results
-      act(() => result.current.handleChange(simulatedInput));
-      act(() =>
-        result.current.searchAndSelectDomain(
-          simulatedEnterKey as React.KeyboardEvent
-        )
-      );
-
-      expect(MOCK_ANIMATE).toHaveBeenCalledWith(leafNode);
-    });
-
-    it("switches semantic domain if given name found", () => {
-      const { result } = renderHook(() => useTreeViewNavigation(testProps));
-
-      const { simulatedInput, simulatedEnterKey } = setupSimulatedInputTest(
-        MockDomain.subdomains[0].name
-      );
-
-      // When testing hooks any call that results in a state change needs to be wrapped in
-      // an act call to avoid warnings and make sure the state change is complete before we test
-      // for the results
-      act(() => result.current.handleChange(simulatedInput));
-      act(() =>
-        result.current.searchAndSelectDomain(
-          simulatedEnterKey as React.KeyboardEvent
-        )
-      );
-
-      expect(MOCK_ANIMATE).toHaveBeenCalledWith(MockDomain.subdomains[0]);
-    });
-
-    it("does not switch semantic domain if given name not found", () => {
-      const TEST: string = "itsatrap";
-      const { result } = renderHook(() => useTreeViewNavigation(testProps));
-      const { simulatedInput, simulatedEnterKey } =
-        setupSimulatedInputTest(TEST);
-      // When testing hooks any call that results in a state change needs to be wrapped in
-      // an act call to avoid warnings and make sure the state change is complete before we test
-      // for the results
-      act(() => result.current.handleChange(simulatedInput));
-      act(() =>
-        result.current.searchAndSelectDomain(
-          simulatedEnterKey as React.KeyboardEvent
-        )
-      );
-
-      expect(MOCK_ANIMATE).toHaveBeenCalledTimes(0);
-    });
-  });
-
   describe("getLeftBrother and getRightBrother", () => {
     it("return undefined when there are no brothers", () => {
-      const { result } = renderHook(() => useTreeViewNavigation(testProps));
+      const { result } = renderHook(() => useTreeNavigation(testProps));
 
       // The top domain (used in testProps) has no brother on either side
       expect(result.current.getLeftBrother(testProps)).toEqual(undefined);
@@ -204,7 +50,7 @@ describe("TreeViewHeader", () => {
     // getBrotherDomain
     it("return the expected brothers", () => {
       const { result } = renderHook(() =>
-        useTreeViewNavigation(upOneWithBrothersProps)
+        useTreeNavigation(upOneWithBrothersProps)
       );
 
       // The top domain (used in testProps) has no brother on either side
@@ -217,68 +63,30 @@ describe("TreeViewHeader", () => {
     });
   });
 
-  // Integration tests, verify the component uses the hooks to achieve the desired UX
-  it("typing non-matching domain search data does not clear input, or attempt to navigate", () => {
-    render(<TreeViewHeader {...upOneWithBrothersProps} />);
-    expect(
-      (screen.getByTestId("testSearch") as HTMLInputElement).value
-    ).toEqual("");
-    userEvent.type(screen.getByTestId("testSearch"), "flibbertigibbet{enter}");
-    expect(
-      (screen.getByTestId("testSearch") as HTMLInputElement).value
-    ).toEqual("flibbertigibbet");
-    // verify that no attempt to switch domains happened
-    expect(MOCK_ANIMATE).toHaveBeenCalledTimes(0);
-  });
+  describe("typing arrow key", () => {
+    it("right arrow moves to right sibling", () => {
+      render(<TreeViewHeader {...upOneWithBrothersProps} />);
+      simulateKey(Key.ArrowRight);
+      expect(MOCK_ANIMATE).toHaveBeenCalledWith(MockDomain.subdomains[2]);
+    });
 
-  it("typing valid domain number navigates and clears input", () => {
-    render(<TreeViewHeader {...upOneWithBrothersProps} />);
-    expect(
-      (screen.getByTestId("testSearch") as HTMLInputElement).value
-    ).toEqual("");
-    userEvent.type(screen.getByTestId("testSearch"), "1.2{enter}");
-    expect(
-      (screen.getByTestId("testSearch") as HTMLInputElement).value
-    ).toEqual("");
-    // verify that we're testing with the matching domain
-    expect(MockDomain.subdomains[2].id).toEqual("1.2");
-    // verify that we would switch to the domain requested
-    expect(MOCK_ANIMATE).toHaveBeenCalledWith(MockDomain.subdomains[2]);
-  });
+    it("left arrow moves to left sibling", () => {
+      render(<TreeViewHeader {...upOneWithBrothersProps} />);
+      simulateKey(Key.ArrowLeft);
+      expect(MOCK_ANIMATE).toHaveBeenCalledWith(MockDomain.subdomains[0]);
+    });
 
-  it("typing right arrow key moves to right sibling", () => {
-    render(<TreeViewHeader {...upOneWithBrothersProps} />);
-    const keyDownHandler = eventListeners.get("keydown");
-    expect(keyDownHandler).not.toBeUndefined();
-    const simulatedArrowKey: Partial<KeyboardEvent> = {
-      key: Key.ArrowRight,
-    };
-    keyDownHandler!.call(null, simulatedArrowKey as Event);
-    // verify that we would switch to the domain requested
-    expect(MOCK_ANIMATE).toHaveBeenCalledWith(MockDomain.subdomains[2]);
-  });
-
-  it("typing left arrow key moves to left sibling", () => {
-    render(<TreeViewHeader {...upOneWithBrothersProps} />);
-    const keyDownHandler = eventListeners.get("keydown");
-    expect(keyDownHandler).not.toBeUndefined();
-    const simulatedArrowKey: Partial<KeyboardEvent> = {
-      key: Key.ArrowLeft,
-    };
-    keyDownHandler!.call(null, simulatedArrowKey as Event);
-    // verify that we would switch to the domain requested
-    expect(MOCK_ANIMATE).toHaveBeenCalledWith(MockDomain.subdomains[0]);
-  });
-
-  it("typing up arrow key moves to parent domain", () => {
-    render(<TreeViewHeader {...upOneWithBrothersProps} />);
-    const keyDownHandler = eventListeners.get("keydown");
-    expect(keyDownHandler).not.toBeUndefined();
-    const simulatedArrowKey: Partial<KeyboardEvent> = {
-      key: Key.ArrowUp,
-    };
-    keyDownHandler!.call(null, simulatedArrowKey as Event);
-    // verify that we would switch to the domain requested
-    expect(MOCK_ANIMATE).toHaveBeenCalledWith(MockDomain);
+    it("up arrow moves to parent domain", () => {
+      render(<TreeViewHeader {...upOneWithBrothersProps} />);
+      simulateKey(Key.ArrowUp);
+      expect(MOCK_ANIMATE).toHaveBeenCalledWith(MockDomain);
+    });
   });
 });
+
+function simulateKey(key: Key) {
+  const keyDownHandler = eventListeners.get("keydown");
+  expect(keyDownHandler).not.toBeUndefined();
+  const simulatedArrowKey: Partial<KeyboardEvent> = { key };
+  keyDownHandler!.call(null, simulatedArrowKey as Event);
+}
